requirements.txt
transformers==4.40.0
torch==2.2.2
fastapi==0.111.0
uvicorn==0.29.0
groq==0.5.0
streamlit==1.33.0
pandas==2.2.2
matplotlib==3.8.4
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.31.0
